requirements.txt
fastapi
uvicorn[standard]
python-multipart
twilio
langchain
langchain-huggingface
langchain_community
langchain-google-genai
faiss-cpu
python-dotenv